+++ library-internal/PostgresqlSyntax/Extras/HeadedMegaparsec.hs
@@ -8,6 +8,7 @@ import Control.Applicative.Combinators hiding (some)
 import Control.Applicative.Combinators.NonEmpty
 import qualified Data.Text as Text
 import HeadedMegaparsec hiding (string)
+import PostgresqlSyntax.Extras.Error (errorBundlePrettyStruct)
 import PostgresqlSyntax.Prelude hiding (bit, expr, filter, head, many, option, some, sortBy, tail, try)
 import Text.Megaparsec (Parsec, Stream, TraversableStream, VisualStream)
 import qualified Text.Megaparsec as Megaparsec
@@ -30,6 +31,14 @@ runParserWithErrorPos p s = case runParser p s of
   Right v -> Right v
   where
     extractor x = (Megaparsec.errorOffset x, Megaparsec.parseErrorPretty x)
+
+-- |
+-- Like 'runParserWithErrorPos' but pairs each error with its
+-- ('Megaparsec.SourcePos') in the parsed message instead of a raw offset.
+runParserWithSourcePosError :: (Show (Megaparsec.Token strm), Show e, Ord e, VisualStream strm, TraversableStream strm, Megaparsec.ShowErrorComponent e) => HeadedParsec e strm a -> strm -> Either (NonEmpty (Megaparsec.SourcePos, String)) a
+runParserWithSourcePosError p s = case runParser p s of
+  Left err -> Left (errorBundlePrettyStruct err)
+  Right v -> Right v
 
 runParser :: (Ord e, VisualStream strm, TraversableStream strm) => HeadedParsec e strm a -> strm -> Either (Megaparsec.ParseErrorBundle strm e) a
 runParser p = Megaparsec.runParser (toParsec p <* Megaparsec.eof) ""